requirements.txt
tensorflow 
streamlit
opencv-python
pillow
pandas
scikit-learn
opendatasets